prog: component Profile input
