+++ src/components/ProfileInput/ProfileInput.module.css
@@ -1,0 +1,71 @@
+.box {
+    display: flex;
+    column-gap: 50px;
+}
+
+.avatar {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    row-gap: 10px;
+}
+
+.foto {
+    width: 170px;
+    height: 170px;
+    border-radius: 50%;
+    background-color: #f0f0f0;
+}
+
+.btnChange {
+    color: #009ee4;
+    font-size: 16px;
+    font-weight: 400;
+    line-height: 150%;
+    background-color: transparent;
+}
+
+.form {
+    width: 630px;
+    display: -webkit-box;
+    display: -ms-flexbox;
+    display: flex;
+    -ms-flex-wrap: wrap;
+    flex-wrap: wrap;
+    row-gap: 20px;
+}
+
+.item {
+    display: flex;
+    flex-direction: column;
+    margin-right: 14px;
+}
+
+.label {
+    color: #c4c4c4;
+    font-size: 16px;
+    font-weight: 500;
+    line-height: 150%;
+    padding-bottom: 4px;
+}
+
+.inputTel,
+.input {
+    color: #000;
+    border: 1px solid rgba(0, 0, 0, 0.2);
+    border-radius: 6px;
+    padding: 13px 19px;
+    width: 300px;
+}
+
+.inputTel {
+    width: 614px;
+}
+
+.inputTel::placeholder,
+.input::placeholder {
+    background-color: transparent;
+    color: rgba(0, 0, 0, 0.3);
+    font-size: 16px;
+    line-height: 24px;
+}
